[Template] Restructure files and folders
