+++ examples/paged-publication-template.html
@@ -16,40 +16,8 @@
         
         <div id="paged-publication"></div>
 
-        <script id="sgn-sdk-paged-publication-viewer-template" type="text/template">
-            <div class="sgn__pp" data-layout-absolute="true">
-                <div class="sgn__header">
-                    <div class="sgn__nav-logo"></div>
-                    <div class="sgn__nav">
-                        <button class="sgn__offer-list">Offers</button>
-                        <button class="sgn__offer-download">Download</button>
-                    </div>
-                </div>
-                <div class="verso">
-                    <div class="verso__scroller">
-                        <div class="sgn-pp__pages"></div>
-                    </div>
-                </div>
-                <div class="sgn-pp__progress">
-                    <div class="sgn-pp-progress__bar"></div>
-                </div>
-                <div class="sgn-pp__progress-label"></div>
-                <a
-                    class="sgn-pp__control"
-                    href="#"
-                    role="button"
-                    data-direction="prev"
-                    >&lsaquo;</a
-                >
-                <a
-                    class="sgn-pp__control"
-                    href="#"
-                    role="button"
-                    data-direction="next"
-                    >&rsaquo;</a
-                >
-            </div>
-        </script>
+        <!-- <script id="sgn-sdk-paged-publication-viewer-template" type="text/template">
+        </script> -->
 
         <!-- 
         <script id="sgn-sdk-paged-publication-viewer-offer-list-template" type="text/template">
