+++ lib/kits/core-ui/components/paged-publication/index.js
@@ -1,3 +1,3 @@
-export {default as PagedContainer} from './paged-container';
+export {default as MainContainer} from './main-container';
 export {default as OfferList} from './offer-list';
 export {default as PublicationDownload} from './publication-download';
