+++ lib/kits/core-ui/components/paged-publication/main-container.js
@@ -1,5 +1,5 @@
 import Mustache from 'mustache';
-import './paged-container.styl';
+import './main-container.styl';
 
 const defaultTemplate = `\
     <div class="sgn__pp" data-layout-absolute="true">
@@ -38,7 +38,7 @@ const defaultTemplate = `\
     </div>\
 `;
 
-export default class PagedContainer {
+export default class MainContainer {
     constructor({template, el}) {
         this.template = template;
         this.el = el;
