+++ lib/kits/core-ui/components/paged-publication/main-container.styl
