+++ lib/kits/core-ui/components/paged-publication/offer-list.js
@@ -1,5 +1,5 @@
 import Mustache from 'mustache';
-import {request} from '../../core';
+import {request} from '../../../core';
 import './offer-list.styl';
 
 const defaultTemplate = `\
