+++ lib/kits/core-ui/components/paged-publication/offer-list.styl
