+++ lib/kits/core-ui/components/paged-publication/publication-download.js
@@ -1,4 +1,4 @@
-import {request} from '../../core';
+import {request} from '../../../core';
 
 export default class PublicationDownload {
     constructor({configs = {}}) {
