+++ lib/kits/core-ui/index.js
@@ -2,6 +2,7 @@ import Gator from '../../../vendor/gator';
 export {default as OfferDetails} from './offer-details';
 export {default as Popover} from './popover';
 export {default as singleChoicePopover} from './single-choice-popover';
+export {default as PagedPublication} from './paged-publication';
 
 export const on = (el, events, selector, callback) =>
     Gator(el).on(events, selector, callback);
