+++ lib/kits/core-ui/paged-publication.js
@@ -1,5 +1,5 @@
 import * as PagedPublicationKit from '../paged-publication';
-import * as components from './components';
+import * as PagedPublicationComponents from './components/paged-publication';
 import {request} from '../core';
 import SGN from '../../sgn-sdk';
 
@@ -8,12 +8,12 @@ export default class PagedPublication {
         this.options = {};
         this.scriptEl = scriptEl;
         this.scriptEls = {
-            pagedContainer: scriptEl.getAttribute(
+            mainContainer: scriptEl.getAttribute(
                 'data-component-paged-publication-viewer-container'
             )
         };
 
-        this.renderPagedContainer();
+        this.renderMainContainer();
     }
 
     async render() {
@@ -25,25 +25,25 @@ export default class PagedPublication {
         this.renderDownload();
     }
 
-    renderPagedContainer() {
-        const pagedContainerEl = document.querySelector(
-            this.scriptEls.pagedContainer
+    renderMainContainer() {
+        const mainContainerEl = document.querySelector(
+            this.scriptEls.mainContainer
         );
         const customTemplate = document.getElementById(
             'sgn-sdk-paged-publication-viewer-template'
         );
-        const pagedContainer = new components.PagedContainer({
+        const mainContainer = new PagedPublicationComponents.MainContainer({
             template: customTemplate,
-            el: pagedContainerEl
+            el: mainContainerEl
         });
-        pagedContainer.render();
+        mainContainer.render();
     }
 
     renderOfferList() {
         const customTemplate = document.getElementById(
             'sgn-sdk-paged-publication-viewer-offer-list-template'
         );
-        const offerList = new components.OfferList({
+        const offerList = new PagedPublicationComponents.OfferList({
             configs: this.options,
             template: customTemplate
         });
@@ -51,7 +51,7 @@ export default class PagedPublication {
     }
 
     renderDownload() {
-        const download = new components.PublicationDownload({
+        const download = new PagedPublicationComponents.PublicationDownload({
             configs: this.options
         });
         download.render();
