--- lib/kits/templates/index.js
@@ -1,1 +1,0 @@
-export {default as PagedPublication} from './paged-publication';
